requirements.txt
alembic==1.8.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
